backend/pyproject.toml
[project]
name = "annotationflow-backend"
version = "0.1.0"
description = "FastAPI backend for building YOLO object-detection datasets from Roboflow Workflow detections."
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "python-multipart==0.0.20",
  "pydantic-settings==2.7.1",
  "Pillow==11.1.0",
  "pillow-heif==0.21.0",
  "inference-sdk>=0.33.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28.1",
  "pytest==8.3.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

